requirements.txt
    # === Core LangGraph & LangChain ===
langgraph>=0.0.30
langchain>=0.1.20
langchain-core>=0.1.50
langchain-community>=0.0.38

# === LLMs & Embeddings ===
openai>=1.3.0
huggingface-hub>=0.21.4
transformers>=4.40.0
tiktoken>=0.5.1

# === Vanna AI ===
vanna==0.7.9

# === Search Tools ===
serpapi>=0.1.5
google-search-results
requests>=2.31.0

# === Document Parsing ===
unstructured>=0.10.25
python-docx>=1.1.0
python-pptx

# === OCR & Image Processing (No sudo required) ===
easyocr
Pillow>=9.0.0
# pdf2image  # Optional: Use only if Poppler is available locally

# === Streamlit UI ===
streamlit>=1.34.0
python-dotenv

# === Vector DB ===
faiss-cpu>=1.7.4

# === Visualization ===
matplotlib==3.10.3
networkx==3.5

# --reading excel
openpyxl
# === Optional Fixes ===
# pydantic<2  # Uncomment only if Pydantic version issues arise

google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client

pypdf


# === Virtual Env Notes ===
# python -m venv venv
# venv\Scripts\activate
# Set-ExecutionPolicy -ExecutionPolicy RemoteSigned -Scope CurrentUser
# venv\Scripts\Activate.ps1

#https://vanna.ai/account/models